requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
blis==0.4.1
boilerpipe==1.3.0.0
boto3==1.12.28
botocore==1.15.28
Brlapi==0.6.6
bs4==0.0.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cycler==0.10.0
cymem==2.0.3
Cython==0.29.15
cytoolz==0.10.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
docutils==0.15.2
en-core-web-sm==2.2.5
entrypoints==0.3
gast==0.3.3
gensim==3.8.1
grammaregex==0.1.3
grpcio==1.27.2
h5py==2.10.0
hdbscan==0.8.24
htmldate==0.6.1
httplib2==0.9.2
idna==2.9
jedi==0.16.0
jellyfish==0.7.2
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.15.1
JPype1==0.7.2
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterthemes==0.20.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
langid==1.1.6
language-selector==0.1
lesscpy==0.14.0
louis==3.5.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
nltk==3.4.5
numpy==1.18.4
olefile==0.45.1
onboard==1.4.1
opencv-python==4.2.0.32
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
ply==3.11
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pyemd==0.5.1
Pygments==2.6.1
pygobject==3.26.1
pyinotify==0.9.6
pyparsing==2.4.6
Pyphen==0.9.5
pyrsistent==0.15.7
pytz==2019.3
pyxdg==0.25
PyYAML==5.3
pyzmq==19.0.0
recordtype==1.3
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
rsa==4.0
s3transfer==0.3.3
scikit-build==0.10.0
scikit-learn==0.23.0
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
six==1.14.0
smart-open==1.10.0
soupsieve==2.0
spacy==2.2.4
srsly==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textacy==0.10.0
textpipeliner==0.3.1
thinc==7.4.0
threadpoolctl==2.0.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
ufw==0.36
ulauncher==5.7.0
uritools==3.0.0
urllib3==1.25.9
urlparse2==1.1.1
uuid==1.30
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.44.0
Werkzeug==1.0.0
wrapt==1.12.1
xkit==0.0.0
YURL==1.0.0
zipp==3.1.0
